Simplified pattern by combining modules
Jira ticket: CAMS-288
