--- user-interface/src/staff-assignment/filters/staffAssignmentFilterStore.ts
@@ -1,10 +1,0 @@
-import { CamsUserReference } from '@common/cams/users';
-
-interface StaffAssignmentFilterStore {
-  officeAssignees: CamsUserReference[];
-  setOfficeAssignees(val: CamsUserReference[]): void;
-  officeAssigneesError: boolean;
-  setOfficeAssigneesError(val: boolean): void;
-}
-
-export type { StaffAssignmentFilterStore };
