--- user-interface/src/staff-assignment/filters/staffAssignmentFilterStoreReact.ts
@@ -1,18 +1,0 @@
-import { CamsUserReference } from '@common/cams/users';
-import { useState } from '../../lib/hooks/UseState';
-
-export function useStaffAssignmentFilterStoreReact() {
-  const [officeAssignees, setOfficeAssignees] = useState<CamsUserReference[]>([]);
-  const [officeAssigneesError, setOfficeAssigneesError] = useState<boolean>(false);
-
-  return {
-    officeAssignees,
-    setOfficeAssignees,
-    officeAssigneesError,
-    setOfficeAssigneesError,
-  };
-}
-
-export default {
-  useStaffAssignmentFilterStoreReact,
-};
