+++ user-interface/src/staff-assignment/filters/staffAssignmentFilterUseCase.test.ts
@@ -1,5 +1,5 @@
 import { ComboOption } from '@/lib/components/combobox/ComboBox';
-import { StaffAssignmentFilterStore } from './staffAssignmentFilterStore';
+import { Store } from './staffAssignmentFilter.types';
 import MockData from '@common/cams/test-utilities/mock-data';
 import staffAssignmentFilterUseCase from './staffAssignmentFilterUseCase';
 import LocalStorage from '@/lib/utils/local-storage';
@@ -20,7 +20,7 @@ describe('staff assignment filter use case tests', () => {
   let setOfficeAssigneesSpy: MockInstance<(val: CamsUserReference[]) => void>;
   let setOfficeAssigneesErrorSpy: MockInstance<(val: boolean) => void>;
   const assignees = MockData.buildArray(MockData.getCamsUserReference, 5);
-  const mockStore: StaffAssignmentFilterStore = {
+  const mockStore: Store = {
     officeAssignees: assignees,
     setOfficeAssignees: vi.fn(),
     officeAssigneesError: false,
